Finished rudimentary data access api and database interaction
Data can only be accessed if it is baked. "Baked" data here means
that it's been parsed and is in the sqlite database. That doesn't
mean that the underlying data has to be still in the folder, though,
as you can point to really any arbitrary sqlite database that's in our
format and it will work just fine. It's yet to be seen what sort of
speed improvements this will give, but in big data sets I'm sure it'll
help a lot. Also, compressing all the parsable data down into a single
database file is pretty rad.