WoW log module added and successful
The Logger module was a little finnicky, and I should probably
document some more about it. Point being, it offers two distinct types
of data for a given range. Namely, 'summary' data which right now only
gives you the number of seconds played for a given date, and 'event'
data which gives you an unparsed string of what happened at a given
time stamp. In the future, I may go back and make the parsing a lot
more elegant, because there's other data that can go in the summary
category and a LOT of information can be parsed from the event logs.